refactor(journal): moved content to separate timetable details components
