+++ src/components/JournalView/JournalTimetables/TimetableDetails/DetailsStops.vue
@@ -1,5 +1,6 @@
 <template>
-  <div class="entry-stops">
+  <div class="details-stops">
+    <h4>TRASA ROZKŁADU:</h4>
     <ul class="stop-list">
       <li v-for="(stop, i) in timetableStops" :key="stop.stopName">
         <span class="stop-label" :data-confirmed="stop.isConfirmed">
@@ -69,28 +70,13 @@
         </span>
       </li>
     </ul>
-
-    <ul class="timetable-path-list" v-if="timetablePathDetails">
-      <li
-        v-for="(pathData, i) in timetablePathDetails"
-        :data-visited="pathData.isVisited"
-        :data-next-visited="
-          i < timetablePathDetails.length - 1 && timetablePathDetails[i + 1].isVisited
-        "
-      >
-        <span v-if="i > 0" class="path-arrow">&gt;</span>
-        <span class="path-arrival" v-if="pathData.arrival">{{ pathData.arrival }}</span>
-        <b class="path-scenery">{{ pathData.sceneryName }}</b>
-        <span class="path-departure" v-if="pathData.departure">{{ pathData.departure }}</span>
-      </li>
-    </ul>
   </div>
 </template>
 
 <script lang="ts">
 import { PropType, defineComponent } from 'vue';
-import dateMixin from '../../../mixins/dateMixin';
-import { API } from '../../../typings/api';
+import dateMixin from '../../../../mixins/dateMixin';
+import { API } from '../../../../typings/api';
 
 interface ITimetableStopDetails {
   stopName: string;
@@ -115,28 +101,6 @@ export default defineComponent({
   },
 
   computed: {
-    timetablePathDetails() {
-      if (!this.timetable.path || this.timetable.path == '') return null;
-
-      return this.timetable.path.split(';').map((pathEl, i) => {
-        const [arrival, name, departure] = pathEl.split(',');
-        const sceneryName = name.split(' ').slice(0, -1).join(' ');
-        const sceneryHash = name.split(' ').pop()?.replace('.sc', '') ?? '';
-        const isVisited = this.timetable.visitedSceneries.includes(sceneryHash);
-
-        return {
-          arrival,
-          sceneryName,
-          sceneryHash,
-          departure,
-          isVisited,
-          isVisitedOffline:
-            !isVisited &&
-            this.timetable.visitedSceneries.includes(`${sceneryName} ${sceneryHash}.sc`)
-        };
-      });
-    },
-
     timetableStops(): ITimetableStopDetails[] {
       const timetable = this.timetable;
 
@@ -145,27 +109,27 @@ export default defineComponent({
       return stopNames.reduce<ITimetableStopDetails[]>((acc, stopName, i, arr) => {
         const arrivalDate =
           i == arr.length - 1
-            ? (timetable.checkpointArrivals.at(i) ?? timetable.endDate)
-            : timetable.checkpointArrivals.at(i);
+            ? (timetable.checkpointArrivals[i] ?? timetable.endDate)
+            : timetable.checkpointArrivals[i];
 
         const scheduledArrivalDate =
           i == arr.length - 1
-            ? (timetable.checkpointArrivalsScheduled.at(i) ?? timetable.scheduledEndDate)
-            : timetable.checkpointArrivalsScheduled.at(i);
+            ? (timetable.checkpointArrivalsScheduled[i] ?? timetable.scheduledEndDate)
+            : timetable.checkpointArrivalsScheduled[i];
 
         const departureDate =
           i == 0
-            ? (timetable.checkpointDepartures.at(i) ?? timetable.beginDate)
-            : timetable.checkpointDepartures.at(i);
+            ? (timetable.checkpointDepartures[i] ?? timetable.beginDate)
+            : timetable.checkpointDepartures[i];
 
         const scheduledDepartureDate =
           i == 0
-            ? (timetable.checkpointDeparturesScheduled.at(i) ?? timetable.scheduledBeginDate)
-            : timetable.checkpointDeparturesScheduled.at(i);
+            ? (timetable.checkpointDeparturesScheduled[i] ?? timetable.scheduledBeginDate)
+            : timetable.checkpointDeparturesScheduled[i];
 
-        const stopTime = Number(timetable.checkpointStopTypes.at(i)?.split(',')[0]) || 0;
-        const stopType = timetable.checkpointStopTypes.at(i)?.split(',').slice(1).join(',') || 'pt';
-        const stopComments = timetable.checkpointComments.at(i) ?? null;
+        const stopTime = Number(timetable.checkpointStopTypes[i]?.split(',')[0]) || 0;
+        const stopType = timetable.checkpointStopTypes[i]?.split(',').slice(1).join(',') || 'pt';
+        const stopComments = timetable.checkpointComments[i] ?? null;
 
         acc.push({
           stopName,
@@ -187,19 +151,19 @@ export default defineComponent({
 </script>
 
 <style lang="scss" scoped>
-@use '../../../styles/badge';
+@use '@/styles/badge';
 
-.entry-stops {
+.details-stops {
   word-wrap: break-word;
   gap: 0.25em;
-  font-size: 0.95em;
 }
 
 .stop-list {
   display: flex;
   flex-wrap: wrap;
   gap: 0.5em;
-  padding: 0.5em 0;
+  margin-top: 0.5em;
+  font-size: 0.95em;
 }
 
 .stop-label {
@@ -251,53 +215,6 @@ export default defineComponent({
   &[data-stop-ph='true'] span,
   &[data-stop-pm='true'] span {
     color: gold;
-  }
-}
-
-.timetable-path-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5em 0;
-  padding: 0.5em 0;
-  color: #ccc;
-
-  li > .path-scenery:first-child,
-  li > .path-arrival:nth-child(2) {
-    border-radius: 0.5em 0 0 0.5em;
-  }
-
-  li > :last-child {
-    border-radius: 0 0.5em 0.5em 0;
-  }
-}
-
-.path-scenery {
-  padding: 0.25em 0.5em;
-  background-color: #303030;
-}
-
-.path-arrival,
-.path-departure {
-  padding: 0.25em;
-  display: inline-block;
-  background-color: #4e4e4e;
-  min-width: 25px;
-  text-align: center;
-}
-
-.path-arrow {
-  padding: 0 0.5em;
-}
-
-.timetable-path-list > li[data-visited='true'] {
-  .path-arrival,
-  .path-scenery,
-  .path-arrow {
-    color: lightgreen;
-  }
-
-  &[data-next-visited='true'] .path-departure {
-    color: lightgreen;
   }
 }
 </style>
