+++ src/styles/_badge.scss
@@ -26,6 +26,22 @@
   }
 }
 
+.specs-badge {
+  margin: 0;
+
+  span:first-child {
+    color: white;
+    background-color: #666;
+    border-radius: 0.25em 0 0 0.25em;
+  }
+
+  span:last-child {
+    color: black;
+    background-color: var(--clr-primary);
+    border-radius: 0 0.25em 0.25em 0;
+  }
+}
+
 .level-badge {
   display: flex;
   justify-content: center;
